# Simpler `PhraseSet::contains_prefix` behavior for ranges