There's not much change from current master, in terms of performance. Since we're using a random sample each time there's going to be some jiggle. The control here is the `exact_contains` bench, which covers `PhraseSet::contains` (a function that wasn't changed).

| benchmark                    | covers                                         | `master` | this PR  |
|------------------------------|------------------------------------------------|----------|----------|
| phrase/exact_contains        | `PhraseSet::contains`                          | 2.3250us | 1.7847us |
| phrase/exact_contains_prefix | `PhraseSet::contains_prefix` (all full words)  | 2.2820us | 2.4375us |
| phrase/range_contains_prefix | `PhraseSet::contains_prefix` (terminal prefix) | 2.5716us | 2.4619us |

That said: the code in this branch is waaaay simpler and I think that's good enough reason to prefer it.