@boblannon 2018-05-31
Still just seeing a lot of apparent noise when comparing runs.

master

```
phrase/exact_contains
			time:   [1.5002 us 1.5283 us 1.5535 us]
phrase/exact_contains_prefix
                        time:   [1.4952 us 1.5301 us 1.5590 us]
phrase/range_contains_prefix
                        time:   [1.8527 us 1.8677 us 1.8832 us]
phrase/range_fst_range
			time:   [2.5146 us 2.5565 us 2.5969 us]
```

pull #26

```
phrase/exact_contains   time:   [1.5739 us 1.5945 us 1.6153 us]
                        change: [+9.7565% +12.781% +15.828%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
phrase/exact_contains_prefix
                        time:   [1.6170 us 1.6422 us 1.6641 us]
                        change: [+1.4625% +5.5872% +9.7369%] (p = 0.01 < 0.05)
                        Performance has regressed.
phrase/range_contains_prefix
                        time:   [1.7529 us 1.7797 us 1.8023 us]
                        change: [-11.583% -8.9233% -6.2535%] (p = 0.00 < 0.05)
                        Performance has improved.
phrase/range_fst_range  time:   [2.8038 us 2.8271 us 2.8473 us]
                        change: [+4.0099% +6.5311% +8.8766%] (p = 0.00 < 0.05)
                        Performance has regressed.
```

master again

```
phrase/exact_contains         time:   [1.6312 us 1.6455 us 1.6598 us]
phrase/exact_contains_prefix  time:   [1.5556 us 1.5657 us 1.5765 us]
phrase/range_contains_prefix  time:   [1.6819 us 1.7124 us 1.7390 us]
phrase/range_fst_range  time:   [2.5849 us 2.6326 us 2.6846 us]

```

pull #25

```
phrase/exact_contains   time:   [1.6019 us 1.6153 us 1.6280 us]
                        change: [-5.1288% -2.1324% +0.7960%] (p = 0.17 > 0.05)
                        No change in performance detected.
phrase/exact_contains_prefix
                        time:   [1.4963 us 1.5197 us 1.5428 us]
                        change: [+1.4050% +4.1455% +6.9894%] (p = 0.00 < 0.05)
                        Performance has regressed.
phrase/range_contains_prefix
                        time:   [1.7457 us 1.7710 us 1.7910 us]
                        change: [-1.7328% +1.8220% +5.7112%] (p = 0.34 > 0.05)
                        No change in performance detected.
phrase/range_fst_range  time:   [2.5678 us 2.6122 us 2.6552 us]
                        change: [-2.3112% +0.3731% +3.0256%] (p = 0.78 > 0.05)
                        No change in performance detected.
```