@boblannon 2018-06-04
after #28 we're doing a better job of building a bench sample that's an approximation of reality. This PR is the clear winner, here.

| bench                        | master                     | #25             | #26             |
|------------------------------|----------------------------|-----------------|-----------------|
| phrase/exact_contains        | [1.5537 **1.5792** 1.6017] | 1.4964 (-5.2%)  | 1.6048          |
| phrase/exact_contains_prefix | [1.4988 **1.5191** 1.5422] | 1.5237          | 1.6062 (+5.7%)  |
| phrase/range_contains_prefix | [2.2230 **2.2380** 2.2522] | 1.3823 (-38.2%) | 1.6506 (-26.2%) |
| phrase/range_fst_range       | [1.2545 **1.2638** 1.2723] | 1.2555          | 1.2291 (-2.7%)  |


<details>

<summary>`master` details</summary>

```
phrase/exact_contains
			time:   [1.5537 us 1.5792 us 1.6017 us]
phrase/exact_contains_prefix
                       	time:   [1.4988 us 1.5191 us 1.5422 us]
found 12 outliers among 100 measurements (12.00%)
  12 (12.00%) low mild
phrase/range_contains_prefix
                       	time:   [2.2230 us 2.2380 us 2.2522 us]
phrase/range_fst_range  time:   [1.2545 us 1.2638 us 1.2723 us]
```

</details>


<details>

<summary>`phrase-prefix-experiment` details</summary>

```
phrase/exact_contains   time:   [1.4606 us 1.4964 us 1.5258 us]
phrase/exact_contains_prefix
                        time:   [1.5086 us 1.5237 us 1.5391 us]
Found 19 outliers among 100 measurements (19.00%)
  1 (1.00%) low severe
  17 (17.00%) low mild
  1 (1.00%) high mild
phrase/range_contains_prefix
                        time:   [1.3651 us 1.3823 us 1.4000 us]
Found 13 outliers among 100 measurements (13.00%)
  10 (10.00%) high mild
  3 (3.00%) high severe
phrase/range_fst_range  time:   [1.2409 us 1.2555 us 1.2686 us]
                        change: [-54.414% -53.483% -52.533%] (p = 0.00 < 0.05)
                        Performance has improved.
```

</details>


<details>

<summary>`phrase-prefix-experiment2` details</summary>

```
file loc is /tmp/fuzzy-phrase-bench/phrase/us_en_latn
phrase/exact_contains   time:   [1.5831 us 1.6048 us 1.6267 us]
                        change: [+12.594% +16.524% +20.435%] (p = 0.00 < 0.05)
                        Performance has regressed.
phrase/exact_contains_prefix
                        time:   [1.5678 us 1.6062 us 1.6382 us]
                        change: [-6.0011% -2.2881% +1.6340%] (p = 0.25 > 0.05)
                        No change in performance detected.
phrase/range_contains_prefix
                        time:   [1.6301 us 1.6506 us 1.6685 us]
                        change: [+16.411% +17.799% +19.220%] (p = 0.00 < 0.05)
                        Performance has regressed.
phrase/range_fst_range  time:   [1.2144 us 1.2291 us 1.2412 us]
                        change: [-3.4295% -1.9361% -0.4250%] (p = 0.02 < 0.05)
                        Change within noise threshold.


```

</details>